Support device properties in expression.
Previously you could reference things such as "hw.totalMemory" in the
datasource's expression, yet if you had a direct property on the device
or component you couldn't use it. This fix solves that problem, and at
the same time allows the more familiar "here.propertyName" and
"device.propertyName" syntaxes to be used.

Fixes ZEN-10648.