Fix bogus calculations when basis counters reset.
See comments on ZEN-20694 for a full explanation.

This fix is a bit of a hack because it assumes that all DERIVE and
COUNTER datapoints are actually counters that may reset or roll over.
This means the calculated performance datasource currently can't be used
on true DERIVE datapoints where the rate of positive or negative change
is what's desired. Since those don't really exist in practice it seemed
like the smarted change to make since doing proper support of DERIVE and
COUNTER would have introduced a lot of potentially dangerous change
close to release.

Fixes ZEN-20694.