Add "Use Basis Interval" to datasource types
By enabling "Use Basis Interval" on Calculated Performance and Datapoint Aggregator datasources, their collection will now attempt to occur as close as possible to the interval of the datapoint they're calculating or aggregating. In the case of multiple basis datapoints with different collection intervals, the shortest will be used.

This new "Use Basis Interval" option will be enabled by default for all existing datasources of these types because this is thought to be the desired functionality in almost all cases.

Fixes ZPS-70.