Change useBasisInterval to use configured interval
Prior to this change useBasisInterval worked in zenpython by changing
the interval of the derived (Calculated Performance & Datapoint
Aggregator) datasource collection based on the timestamps of the
previous two collections of their basis datapoints. In practice (see
ZPS-2077) this seems to result in irregular collection intervals, and
not a terribly good match to the basis intervals.

This change moves the implementation of useBasisInterval from the
collector (zenpython) into the hub config service (PythonConfig). Now
instead of looking at the difference between the previous two collection
intervals of the basis datasources, we only look at the configured
cycletime (or getCycleTime(context) when available) of the basis
datasources. This results in a perfectly predictable collection interval
for the derived datasources at a slightly higher cost when building
zenpython's configuration.

I profiled how long it takes to build the zenpython configuration for
our "ucs1" on UCS-PM 2.5.0 before and after this change.

Before this change:

    4629860 function calls (4438746 primitive calls) in 27.950 seconds

After this change:

    5657494 function calls (5398178 primitive calls) in 40.939 seconds

I haven't found a way to optimize this any further, and I believe the
additional cost is worth it. UCS-PM with large UCS devices is a rather
extreme example because the derived datapoints are nested many levels
deep, and there's a higher than normal cost to even calculate the
cycletime of the basis datapoints.

Note that for UCS this change relies on a change going into CiscoUCS
2.6.2 that implements UCSDataSource.getCycleTime to provide the true
collection interval of all UCS datasources that would be the basis
datasources in UCS-PM.

Refs ZPS-2077.