Fetch data for datasources with interval > 600
A couple of places had hard-coded only looking for the last basis
datapoint up to 600 second in the past. This doesn't work for
datasources collected less frequently than every 600 seconds.

These changes using the convention of looking up to
(datasource.cycletime * 5) seconds in the past for basis datapoints.

Fixes ZPS-2077.